remove urql client from indexer common
