+++ packages/indexer-cli/src/client.ts
@@ -1,11 +1,6 @@
 import { createClient } from '@urql/core'
 import fetch from 'isomorphic-fetch'
-import { IndexerManagementClient } from '@graphprotocol/indexer-common'
 
-export const createIndexerManagementClient = async ({
-  url,
-}: {
-  url: string
-}): Promise<IndexerManagementClient> => {
-  return createClient({ url, fetch }) as unknown as IndexerManagementClient
+export const createIndexerManagementClient = async ({ url }: { url: string }) => {
+  return createClient({ url, fetch })
 }
