+++ packages/indexer-common/package.json
@@ -30,8 +30,6 @@
     "@types/lodash.clonedeep": "^4.5.7",
     "@types/lodash.intersection": "^4.4.7",
     "@types/lodash.xor": "^4.5.7",
-    "@urql/core": "2.4.4",
-    "@urql/exchange-execute": "1.2.2",
     "axios": "1.6.2",
     "body-parser": "1.20.2",
     "cors": "2.8.5",
@@ -86,8 +84,7 @@
   "resolutions": {
     "ethers": "5.7.0",
     "sequelize": "6.33.0",
-    "@ethersproject/bignumber": "5.7.0",
-    "@urql/exchange-execute/@urql/core": "2.4.4"
+    "@ethersproject/bignumber": "5.7.0"
   },
   "babel": {
     "presets": [],
