+++ packages/indexer-common/src/schema/indexer-management/resolvers/Mutation/deleteIndexingRule.ts
@@ -1,7 +1,7 @@
 import { processIdentifier } from '../../../../subgraphs'
 import type { MutationResolvers } from './../../../types.generated'
 import { validateNetworkIdentifier } from '../../../../parsers/validators'
-import { resetGlobalRule } from '../../../../indexer-management/resolvers/indexing-rules'
+import { resetGlobalRule } from '../utils'
 
 export const deleteIndexingRule: NonNullable<
   MutationResolvers['deleteIndexingRule']
