+++ packages/indexer-common/src/schema/indexer-management/resolvers/Mutation/reallocateAllocation.ts
@@ -1,4 +1,4 @@
-import { extractNetwork } from '../../../../indexer-management/resolvers/utils'
+import { extractNetwork } from '../utils'
 import {
   IdentifierType,
   IndexingDecisionBasis,
