+++ packages/indexer-common/src/schema/indexer-management/resolvers/Query/indexerRegistration.ts
@@ -1,4 +1,4 @@
-import { extractNetwork } from '../../../../indexer-management/resolvers/utils'
+import { extractNetwork } from '../utils'
 import geohash from 'ngeohash'
 import type { QueryResolvers } from './../../../types.generated'
