+++ packages/indexer-service/package.json
@@ -35,6 +35,7 @@
     "@graphprotocol/common-ts": "2.0.9",
     "@graphprotocol/indexer-common": "^0.21.2",
     "@graphprotocol/indexer-native": "0.20.11",
+    "@graphql-tools/executor-http": "^1.0.9",
     "@graphql-tools/load": "8.0.0",
     "@graphql-tools/url-loader": "8.0.0",
     "@graphql-tools/wrap": "10.0.1",
