+++ packages/indexer-service/src/commands/start.ts
@@ -17,7 +17,7 @@ import {
   toAddress,
 } from '@graphprotocol/common-ts'
 import {
-  createIndexerManagementClient,
+  createIndexerManagementYogaClient,
   defineIndexerManagementModels,
   defineQueryFeeModels,
   indexerError,
@@ -456,7 +456,7 @@ export default {
       queryTimingLogs: argv.queryTimingLogs,
     })
 
-    const indexerManagementClient = await createIndexerManagementClient({
+    const indexerManagementClient = await createIndexerManagementYogaClient({
       models,
       graphNode,
       indexNodeIDs: ['node_1'], // This is just a dummy since the indexer-service doesn't manage deployments,
